TST/FIX: another attempt to fix `itt` being reported as 0, improve results
Reviewing older versions reveal the exponent was 0.333 instead of
the expression (1/3.) so revised to match.

To distinguish between "not able to calculate" and "conditions not
appropriate" return -1 if zeta is out of range and NAN for other errors.

Return whole values only to emphasize uncertainty.

Refactor substantially, but might not work correctly: `IIF` instruction
has been observed misbehaving in subroutines when interacting with
variables accessible through custom menus. Since this is function and
no menu vars are available, it's deemed safe enough to try out.